Fixes of the custom step settings
- Switched scoreStep from float to double
- Rename from BottomSheetPreference to ScoreStepsPreferenceSheet

+++ feature/settings/src/main/java/com/axiel7/anihyou/feature/settings/SettingsEvent.kt
@@ -63,7 +63,8 @@ interface SettingsEvent : UiEvent {
     fun setStaffNameLanguage(value: UserStaffNameLanguage)
 
     fun setScoreFormat(value: ScoreFormat)
-    fun setScoreStep(value: Float)
+
+    fun setScoreStep(value: Double)
 
     fun setScoreStep(value: Double)
